si2168: Add ts bus control, turn off bus on sleep
Includes a function to set TS MODE property os si2168. The function
either disables the TS output bus, or sets mode to config option.

After setting up the frontend the TS bus is enabled. When going
to sleep the TS bus is tri-stated, this makes the driver
compatible with multiple frontend usage.

Signed-off-by: Brad Love <brad@nextdimension.cc>